Extract interrupt handling to interrupts.go
